MPS/p1/requirements.txt
SpeechRecognition
numpy
matplotlib
opencv-python
pillow
easygui
pyaudio